requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.1.0
# Optional: For AWS S3 storage (install when needed)
# django-storages==1.14.2
# boto3==1.34.0

